requirements.txt
yfinance==0.2.35
python-telegram-bot==21.0.1
apscheduler==3.10.4
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
